Add keepWebhook flag when deleting entities

+++ runner/organizations.go
@@ -124,7 +124,7 @@ func (r *Runner) GetOrganizationByID(ctx context.Context, orgID string) (params.
 	return org, nil
 }
 
-func (r *Runner) DeleteOrganization(ctx context.Context, orgID string) error {
+func (r *Runner) DeleteOrganization(ctx context.Context, orgID string, keepWebhook bool) error {
 	if !auth.IsAdmin(ctx) {
 		return runnerErrors.ErrUnauthorized
 	}
@@ -148,14 +148,16 @@ func (r *Runner) DeleteOrganization(ctx context.Context, orgID string) error {
 		return runnerErrors.NewBadRequestError("org has pools defined (%s)", strings.Join(poolIds, ", "))
 	}
 
-	poolMgr, err := r.poolManagerCtrl.GetOrgPoolManager(org)
-	if err != nil {
-		return errors.Wrap(err, "fetching pool manager")
-	}
+	if !keepWebhook {
+		poolMgr, err := r.poolManagerCtrl.GetOrgPoolManager(org)
+		if err != nil {
+			return errors.Wrap(err, "fetching pool manager")
+		}
 
-	if err := poolMgr.UninstallWebhook(ctx); err != nil {
-		// TODO(gabriel-samfira): Should we error out here?
-		log.Printf("failed to uninstall webhook: %s", err)
+		if err := poolMgr.UninstallWebhook(ctx); err != nil {
+			// TODO(gabriel-samfira): Should we error out here?
+			log.Printf("failed to uninstall webhook: %s", err)
+		}
 	}
 
 	if err := r.poolManagerCtrl.DeleteOrgPoolManager(org); err != nil {
